+++ examples/tiny-url/src/main/scala/com/twitter/tiny/modules/ServicesModule.scala
@@ -1,10 +1,8 @@
 package com.twitter.tiny.modules
 
-import com.google.inject.{Provides, Singleton}
 import com.twitter.inject.TwitterModule
-import com.twitter.tiny.services.{Counter, UrlShortenerService}
+import com.twitter.tiny.services.UrlShortenerService
 import com.twitter.tiny.services.impl.RedisUrlShortenerService
-import redis.clients.jedis.{Jedis => JedisClient}
 
 object ServicesModule extends TwitterModule {