+++ http/src/main/scala/com/twitter/finatra/http/internal/marshalling/FinatraFileUpload.scala
@@ -2,7 +2,7 @@ package com.twitter.finatra.http.internal.marshalling
 
 import com.twitter.finagle.http.{Method, Request}
 import com.twitter.finatra.http.fileupload.MultipartItem
-import org.apache.commons.fileupload.{FileItemFactory, FileItemHeaders, FileItemIterator, FileUploadBase}
+import org.apache.commons.fileupload.{FileItemFactory, FileItemIterator, FileUploadBase}
 import org.apache.commons.io.IOUtils
 import scala.collection.mutable