+++ thrift/src/test/scala/com/twitter/finatra/thrift/tests/doeverything/DoEverythingThriftServer.scala
@@ -1,13 +1,14 @@
 package com.twitter.finatra.thrift.tests.doeverything
 
+import com.twitter.finagle.ThriftMux
+import com.twitter.finagle.tracing.NullTracer
 import com.twitter.finatra.thrift.filters._
 import com.twitter.finatra.thrift.modules.ClientIdWhitelistModule
 import com.twitter.finatra.thrift.routing.ThriftRouter
 import com.twitter.finatra.thrift.tests.doeverything.controllers.DoEverythingThriftController
 import com.twitter.finatra.thrift.tests.doeverything.filters.ExceptionTranslationFilter
 import com.twitter.finatra.thrift.{ThriftFilter, ThriftServer}
-import com.twitter.inject.Logging
-import com.twitter.inject.utils.Handler
+import com.twitter.util.NullMonitor
 
 object DoEverythingThriftServerMain extends DoEverythingThriftServer
 
@@ -18,6 +19,12 @@ class DoEverythingThriftServer extends ThriftServer {
 
   override val modules = Seq(
     ClientIdWhitelistModule)
+
+  override protected def configureThriftServer(server: ThriftMux.Server): ThriftMux.Server = {
+    server
+      .withMonitor(NullMonitor)
+      .withTracer(NullTracer)
+  }
 
   override def configureThrift(router: ThriftRouter): Unit = {
     router
@@ -37,13 +44,3 @@ class DoEverythingThriftServer extends ThriftServer {
   }
 }
 
-class DoEverythingThriftWarmupHandler
-  extends Handler
-  with Logging {
-
-  override def handle() = {
-    // TODO: implement ThriftWarmup
-    // warm up thrift service(s) here
-    info("Warm-up done.")
-  }
-}